requirements.txt
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.1
matplotlib==3.7.5
Pillow==9.5.0  
opencv-python==4.9.0 
torch==2.1.2+cu118
albumentations==1.4.0
segmentation-models-pytorch==0.3.3
monai==1.3.2
scikit-learn==1.2.2